fix: Update configs and handlers

+++ internal/groundcontrol/server/handleResponse.go
@@ -13,7 +13,7 @@ func (e *AppError) Error() string {
 
 // write JSON error response with given status code and message.
 func WriteJSONError(w http.ResponseWriter, message string, statusCode int) {
-	respBytes, err := json.Marshal(AppError{Message: message, Code: int64(statusCode)})
+	respBytes, err := json.Marshal(AppError{Message: message, Code: statusCode})
 	if err != nil {
 		log.Printf("Failed to marshal JSON error response: %v", err)
 		w.Header().Set("Content-Type", "application/json")
@@ -51,7 +51,7 @@ func WriteJSONResponse(w http.ResponseWriter, statusCode int, data any) {
 func HandleAppError(w http.ResponseWriter, err error) {
 	var appErr *AppError
 	if errors.As(err, &appErr) {
-		WriteJSONResponse(w, int(appErr.Code), appErr)
+		WriteJSONResponse(w, appErr.Code, appErr)
 	} else {
 		WriteJSONResponse(w, http.StatusInternalServerError, &AppError{
 			Message: "Internal Server Error",
